% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download Dataset}
\usage{
download_dataset(title, download_dir = "datasets", format = NULL)
}
\arguments{
\item{title}{Character. Title of the dataset to download.}

\item{download_dir}{Character. Directory to save the downloaded dataset at, defaults to "datasets".}

\item{format}{Character. Format of the dataset to download.}
}
\value{
The demanded dataset in the demanded path.
}
\description{
Downloads a dataset from the Tunisian data catalog API (data.gov.tn).
}
\examples{
\donttest{
try({
  download_dataset("DONNÉES CLIMATIQUES STATION Tozeur AVFA", format = "csv")
})
}

}
