% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summaries.R
\name{get_authors}
\alias{get_authors}
\title{List Authors/Organizations}
\usage{
get_authors(limit = 20)
}
\arguments{
\item{limit}{Integer. Maximum number of organizations to return, defaults to 10.}
}
\value{
A data frame with columns for id, name, dataset_count, and description.
}
\description{
Retrieves organizations data from the Tunisian data catalog API (data.gov.tn).
(Caution - this function may be slow if you choose to retrieve all organizations)
}
