% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_keywords}
\alias{get_keywords}
\title{List Dataset Keywords/Tags}
\usage{
get_keywords(limit = 10, query = NULL)
}
\arguments{
\item{limit}{Integer. Maximum number of tags to return (default: 10).}

\item{query}{Character. Optional search string to filter tags.}
}
\value{
A data frame of keywords/tags with counts.
}
\description{
Retrieves a list of unique keywords/tags from the Tunisian data catalog API.
}
\examples{
\donttest{
try({
keywords <- get_keywords(limit = 10)
})
}

}
