% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_themes}
\alias{get_themes}
\title{Summarize Dataset Themes}
\usage{
get_themes()
}
\value{
A tibble (data frame) with two columns:
\describe{
\item{theme}{Character. Name of the theme/group.}
\item{dataset_count}{Numeric. Number of datasets in the theme.}
}
}
\description{
Fetches and summarizes themes (groups) alongside the number of datasets in
each theme from the Tunisian data catalog API (data.gov.tn).
}
\examples{
try({
  themes_summary <- get_themes()
  head(themes_summary)
})

}
